Enum dqcsim::bindings::dqcs_handle_type_t

#[repr(C)]
pub enum dqcs_handle_type_t {
    DQCS_HTYPE_INVALID,
    DQCS_HTYPE_ARB_DATA,
    DQCS_HTYPE_ARB_CMD,
    DQCS_HTYPE_ARB_CMD_QUEUE,
    DQCS_HTYPE_QUBIT_SET,
    DQCS_HTYPE_GATE,
    DQCS_HTYPE_MEAS,
    DQCS_HTYPE_MEAS_SET,
    DQCS_HTYPE_MATRIX,
    DQCS_HTYPE_GATE_MAP,
    DQCS_HTYPE_FRONT_PROCESS_CONFIG,
    DQCS_HTYPE_OPER_PROCESS_CONFIG,
    DQCS_HTYPE_BACK_PROCESS_CONFIG,
    DQCS_HTYPE_FRONT_THREAD_CONFIG,
    DQCS_HTYPE_OPER_THREAD_CONFIG,
    DQCS_HTYPE_BACK_THREAD_CONFIG,
    DQCS_HTYPE_SIM_CONFIG,
    DQCS_HTYPE_SIM,
    DQCS_HTYPE_FRONT_DEF,
    DQCS_HTYPE_OPER_DEF,
    DQCS_HTYPE_BACK_DEF,
    DQCS_HTYPE_PLUGIN_JOIN,
}

Enumeration of types that can be associated with a handle.

Variants

DQCS_HTYPE_INVALID

Indicates that the given handle is invalid.

This indicates one of the following:

The handle value is invalid (zero or negative).

The handle has not been used yet.

The object associated with the handle was deleted.

DQCS_HTYPE_ARB_DATA

Indicates that the given handle belongs to an ArbData object.

This means that the handle supports the handle and arb interfaces.

DQCS_HTYPE_ARB_CMD

Indicates that the given handle belongs to an ArbCmd object.

This means that the handle supports the handle, arb, and cmd interfaces.

DQCS_HTYPE_ARB_CMD_QUEUE

Indicates that the given handle belongs to a queue of ArbCmd object.

This means that the handle supports the handle, arb, cmd, and cq interfaces.

DQCS_HTYPE_QUBIT_SET

Indicates that the given handle belongs to a set of qubit references.

This means that the handle supports the handle and qbset interfaces.

DQCS_HTYPE_GATE

Indicates that the given handle belongs to a quantum gate description.

This means that the handle supports the handle, gate, and arb interfaces.

DQCS_HTYPE_MEAS

Indicates that the given handle belongs to a qubit measurement result.

This means that the handle supports the handle, meas, and arb interfaces. It can also be used in place of a qubit measurement result set by functions that consume the object.

DQCS_HTYPE_MEAS_SET

Indicates that the given handle belongs to a set of qubit measurement results.

This means that the handle supports the handle and mset interfaces.

DQCS_HTYPE_MATRIX

Indicates that the given handle belongs to a matrix. >

This means that the handle supports the handle and mat interfaces.

DQCS_HTYPE_GATE_MAP

Indicates that the given handle belongs to a gate map. >

This means that the handle supports the handle and gm interfaces.

DQCS_HTYPE_FRONT_PROCESS_CONFIG

Indicates that the given handle belongs to a frontend plugin process configuration object.

This means that the handle supports the handle, pcfg, and xcfg interfaces.

DQCS_HTYPE_OPER_PROCESS_CONFIG

Indicates that the given handle belongs to an operator plugin process configuration object.

This means that the handle supports the handle, pcfg, and xcfg interfaces.

DQCS_HTYPE_BACK_PROCESS_CONFIG

Indicates that the given handle belongs to a backend plugin process configuration object.

This means that the handle supports the handle, pcfg, and xcfg interfaces.

DQCS_HTYPE_FRONT_THREAD_CONFIG

Indicates that the given handle belongs to a frontend plugin thread configuration object.

This means that the handle supports the handle, tcfg, and xcfg interfaces.

DQCS_HTYPE_OPER_THREAD_CONFIG

Indicates that the given handle belongs to an operator plugin thread configuration object.

This means that the handle supports the handle, tcfg, and xcfg interfaces.

DQCS_HTYPE_BACK_THREAD_CONFIG

Indicates that the given handle belongs to a backend plugin thread configuration object.

This means that the handle supports the handle, tcfg, and xcfg interfaces.

DQCS_HTYPE_SIM_CONFIG

Indicates that the given handle belongs to a simulator configuration object.

This means that the handle supports the handle and scfg interfaces.

DQCS_HTYPE_SIM

Indicates that the given handle belongs to a simulator instance.

This means that the handle supports the handle and sim interfaces.

DQCS_HTYPE_FRONT_DEF

Indicates that the given handle belongs to a frontend plugin definition object.

This means that the handle supports the handle and pdef interfaces.

DQCS_HTYPE_OPER_DEF

Indicates that the given handle belongs to an operator plugin definition object.

This means that the handle supports the handle and pdef interfaces.

DQCS_HTYPE_BACK_DEF

Indicates that the given handle belongs to a backend plugin definition object.

This means that the handle supports the handle and pdef interfaces.

DQCS_HTYPE_PLUGIN_JOIN

Indicates that the given handle belongs to a plugin thread join handle.

This means that the handle supports the handle and pjoin interfaces.