Function dqcsim::bindings::dqcs_pcfg_init_cmd

#[no_mangle]
pub extern "C" fn dqcs_pcfg_init_cmd(
    pcfg: dqcs_handle_t, 
    cmd: dqcs_handle_t
) -> dqcs_return_t

Appends an ArbCmd to the list of initialization commands of a plugin process.

The ArbCmd handle is consumed by this function, and is thus invalidated, if and only if it is successful.