Function dqcsim::bindings::dqcs_pcfg_new

#[no_mangle]
pub extern "C" fn dqcs_pcfg_new(
    typ: dqcs_plugin_type_t, 
    name: *const c_char, 
    spec: *const c_char
) -> dqcs_handle_t

Creates a new plugin process configuration object using sugared syntax.

typ specifies the type of plugin. name specifies the name used to refer to the plugin later, which much be unique within a simulation; if it is empty or NULL, auto-naming will be performed: "front" for the frontend, "oper<i>" for the operators (indices starting at 1 from frontend to backend), and "back" for the backend. spec specifies which plugin to use, using the same syntax that the dqcsim command line interface uses.