Function dqcsim::bindings::dqcs_plugin_random_f64

#[no_mangle]pub extern "C" fn dqcs_plugin_random_f64(
    plugin: dqcs_plugin_state_t
) -> c_double

Generates a random floating point number using the simulator random seed.

The generated numbers are uniformly distributed in the range [0,1>.

This function only fails if the plugin handle is invalid, in which case it returns 0. Of course, 0 is also a valid (if rare) random return value.