Function dqcsim::bindings::dqcs_scfg_stderr_verbosity_get

#[no_mangle]pub extern "C" fn dqcs_scfg_stderr_verbosity_get(
    scfg: dqcs_handle_t
) -> dqcs_loglevel_t

Returns the configured stderr sink verbosity for a simulation.

That is, the minimum loglevel that a messages needs to have for it to be printed to stderr.