Function dqcsim::bindings::dqcs_sim_arb

#[no_mangle]pub extern "C" fn dqcs_sim_arb(
    sim: dqcs_handle_t, 
    name: *const c_char, 
    cmd: dqcs_handle_t
) -> dqcs_handle_t

Sends an ArbCmd message to one of the plugins, referenced by name.

ArbCmds are executed immediately after yielding to the simulator, so all pending asynchronous calls are flushed and executed before the ArbCmd.

When this succeeds, the received data is returned in the form of a new handle. When it fails, 0 is returned.

The ArbCmd handle is consumed if and only if the API call succeeds.