Module dqcsim::common::channel

Channel abstractions.

Defines a Channel trait based on a Sender and Receiver trait combination to abstract over different channel types.

Traits

Channel
Channel abstraction.

Receiver
Receiver side of a channel.

Sender
Sender side of a channel.

Type Definitions

CrossbeamChannel

DownstreamChannel

IpcChannel

PluginChannel

SimulatorChannel

UpstreamChannel