Struct dqcsim::common::converter::ConverterMap

pub struct ConverterMap<'c, K, I, O> where
    K: Eq + Hash + Clone,
    I: Eq + Hash + Clone,
    O: Clone,  { /* fields omitted */ }

K: user-defined key for identifying which converter to use I: detector input = constructor output O: detector output = constructor input

Implementations

impl<'c, K, I, O> ConverterMap<'c, K, I, O> where
K: Eq + Hash + Clone,
I: Eq + Hash + Clone,
O: Clone,

pub fn new(cache_key_generator: Option<Box<dyn Fn(&I) -> I>>) -> Self

Constructs a new empty ConverterMap.

pub fn push(
&mut self,
key: impl Into<K>,
converter: Box<dyn Converter<Input = I, Output = O> + 'c>
)

Appends a Converter with the specified key to the back of the collection of Detectors in this map.

pub fn insert(
&mut self,
index: usize,
key: impl Into<K>,
converter: Box<dyn Converter<Input = I, Output = O> + 'c>
)

Inserts a Converter at position index within the collection of Detectors in this map, with the specified key associated to the inserted Converter.

pub fn with(
self,
key: impl Into<K>,
converter: Box<dyn Converter<Input = I, Output = O> + 'c>
) -> Self

Appends the specified Converter with the corresponding specified key to the collection and returns the updated DetectorMap.

pub fn clear_cache(&self)

Clears the cache.

pub fn len(&self) -> usize

Returns the number of Detectors in the collection.

pub fn is_empty(&self) -> bool

Returns true if the collection contains no Detectors.