Trait dqcsim::common::converter::FromArb

pub trait FromArb where
    Self: Sized, {
    fn from_arb(src: &mut ArbData) -> Result<Self>;
}

A type that can be constructed from (part of) an ArbData object.

Any data not consumed by the construction stays in the ArbData object for further interpretation.

Required methods

fn from_arb(src: &mut ArbData) -> Result<Self>

Construct from the given ArbData, taking the parameters used for construction out of the data object.

Loading content...