Trait dqcsim::common::converter::ToArb

pub trait ToArb {
    fn to_arb(self, dest: &mut ArbData);
}

A type that can be converted into (part of) an ArbData object.

This is the reverse of FromArb. Generally, constructing an arb from a number of ToArb objects in a certain order should make the objects recoverable using FromArb in the reverse order, assuming all involved types implement both directions.

Required methods

fn to_arb(self, dest: &mut ArbData)

Convert self to ArbData parameters and add them to the data object.

Loading content...