Module dqcsim::common::error

Error-handling.

Types for error-handling in this crate, based on the failure crate.

Error is the wrapper which implements Fail and containers the inner ErrorKind and its Context.

Structs

Error
Error type for this crate.

Enums

ErrorKind

Traits

ResultExt
Re-export the ResultExt trait which adds the [Context] methods to [Result].

Functions

err
Shorthand for producing an error that does not fit in any of the ErrorKind classes.

inv_arg
Shorthand for producing an invalid argument error.

inv_op
Shorthand for producing an invalid operation error.

log_err
Shorthand for producing a LogError.

oe_err
err() but for or_else() functions.

oe_inv_arg
Shorthand for producing an invalid argument error in an or_else function.

oe_inv_op
Shorthand for producing an invalid operation error in an or_else function.

oe_log_err
Shorthand for producing a LogError in an or_else function.

Type Definitions

Result
Internal [Result] type which uses the crate's Error type.