Module dqcsim::common::gates

Helper functions for common quantum Gates

Enums

Gates
Enumeration of predefined quantum gates.

Functions

cnot_gate
Returns a CNOT gate.

fredkin_gate
Returns a Fredking gate.

h_gate
Returns a Hadamard gate.

i_gate
Returns an I gate.

mx90_gate
Returns a negative 90-degree X gate.

my90_gate
Returns a negative 90-degree Y gate.

mz90_gate
Returns a negative 90-degree Z gate.

rx_gate
Returns an arbitrary X rotation gate. Theta is the rotation angle in radians.

ry_gate
Returns an arbitrary Y rotation gate. Theta is the rotation angle in radians.

rz_gate
Returns an arbitrary Y rotation gate. Theta is the rotation angle in radians.

s_gate
Returns an S gate.

sdag_gate
Returns an S-dagger gate.

sqswap_gate
Returns a square-root-of-swap gate on provided target qubits a and b.

swap_gate
Returns a swap gate on provided target qubits a and b.

t_gate
Returns a T gate.

tdag_gate
Returns a T-dagger gate.

toffoli_gate
Returns a Toffoli gate.

x90_gate
Returns a 90-degree X gate.

x_gate
Returns an X gate.

y90_gate
Returns a 90-degree Y gate.

y_gate
Returns a Y gate.

z90_gate
Returns a 90-degree Z gate.

z_gate
Returns a Z gate.