Enum dqcsim::common::gates::BoundUnitaryGate

pub enum BoundUnitaryGate<'matrix, 'qref> {
    I(QubitRef),
    X(QubitRef),
    Y(QubitRef),
    Z(QubitRef),
    H(QubitRef),
    S(QubitRef),
    SDAG(QubitRef),
    T(QubitRef),
    TDAG(QubitRef),
    RX90(QubitRef),
    RXM90(QubitRef),
    RX180(QubitRef),
    RY90(QubitRef),
    RYM90(QubitRef),
    RY180(QubitRef),
    RZ90(QubitRef),
    RZM90(QubitRef),
    RZ180(QubitRef),
    RX(f64, QubitRef),
    RY(f64, QubitRef),
    RZ(f64, QubitRef),
    Phase(f64, QubitRef),
    PhaseK(u64, QubitRef),
    R(f64, f64, f64, QubitRef),
    SWAP(QubitRef, QubitRef),
    SQSWAP(QubitRef, QubitRef),
    U(&'matrix Matrix, &'qref [QubitRef]),
}

A bound gate type.

The variants in this type specify all parameters and qubits targets to determine the behaviour of the gate.

Variants

I(QubitRef)

Identity with specified qubit target.

X(QubitRef)

Pauli-X with specified qubit target.

Y(QubitRef)

Pauli-Y with specified qubit target.

Z(QubitRef)

Pauli-Z with specified qubit target.

H(QubitRef)

Hadamard with specified qubit target.

S(QubitRef)

Phase with specified qubit target.

SDAG(QubitRef)

S† (conjugate transpose of S) with specified qubit target.

T(QubitRef)

T with specified qubit target.

TDAG(QubitRef)

T† (conjugate transpose of T) with specified qubit target.

RX90(QubitRef)

90 degree rotation around X-axis with specified qubit target.

RXM90(QubitRef)

minus 90 degree rotation around X-axi with specified qubit target.

RX180(QubitRef)

180 degree rotation around X-axis with specified qubit target.

RY90(QubitRef)

90 degree rotation around Y-axis with specified qubit target.

RYM90(QubitRef)

minus 90 degree rotation around Y-axis with specified qubit target.

RY180(QubitRef)

180 degree rotation around Y-axis with specified qubit target.

RZ90(QubitRef)

90 degree rotation around Y-axis with specified qubit target.

RZM90(QubitRef)

minus 90 degree rotation around Z-axis with specified qubit target.

RZ180(QubitRef)

180 degree rotation around Z-axis with specified qubit target.

RX(f64, QubitRef)

Arbitrary rotation around X-axis with specified angle (θ) and qubit target.

RY(f64, QubitRef)

Arbitrary rotation around Y-axis with specified angle (θ) and qubit target.

RZ(f64, QubitRef)

Arbitrary rotation around Z-axis with specified angle (θ) and qubit target.

Phase(f64, QubitRef)

Arbitrary rotation around Z-axis, global phase chosen such that it works as a submatrix for controlled phase operations.

PhaseK(u64, QubitRef)

Same as Phase, but with θ = π/2^k​.

R(f64, f64, f64, QubitRef)

Arbitrary rotation around X-, Y- and Z-axis with specified angles (θ, φ, λ) and qubit target.

SWAP(QubitRef, QubitRef)

Swap with specified qubit targets.

SQSWAP(QubitRef, QubitRef)

Square root of Swap with specified qubit targets.

U(&'matrix Matrix, &'qref [QubitRef])

Abstract unitary gate with a reference to specified unitary matrix and qubit targets.