Enum dqcsim::common::gates::UnitaryGateType

pub enum UnitaryGateType {
    I,
    X,
    Y,
    Z,
    H,
    S,
    SDAG,
    T,
    TDAG,
    RX90,
    RXM90,
    RX180,
    RY90,
    RYM90,
    RY180,
    RZ90,
    RZM90,
    RZ180,
    RX,
    RY,
    RZ,
    Phase,
    PhaseK,
    R,
    SWAP,
    SQSWAP,
    U(usize),
}

An abstract gate type.

The variants in this type carry no additional parameters or information about the target qubits, with the exception of the UnitaryGateType::U variant that encodes an abstract Unitary gate with the number of qubits involved specified.

Variants

I

Identity.

X

Pauli-X.

Y

Pauli-Y.

Z

Pauli-Z.

H

Hadamard.

S

Phase.

SDAG

S† (conjugate transpose of S).

T

T.

TDAG

T† (conjugate transpose of T).

RX90

90 degree rotation around X-axis.

RXM90

minus 90 degree rotation around X-axis.

RX180

180 degree rotation around X-axis.

RY90

90 degree rotation around Y-axis.

RYM90

minus 90 degree rotation around Y-axis.

RY180

180 degree rotation around Y-axis.

RZ90

90 degree rotation around Y-axis.

RZM90

minus 90 degree rotation around Z-axis.

RZ180

180 degree rotation around Z-axis.

RX

Arbitrary rotation around X-axis.

RY

Arbitrary rotation around Y-axis.

RZ

Arbitrary rotation around Z-axis.

Phase

Arbitrary rotation around Z-axis, global phase chosen such that it works as a submatrix for controlled phase operations.

PhaseK

Same as Phase, but with θ = π/2^k​.

R

Arbitrary rotation around X-, Y- and Z-axis.

SWAP

Swap.

SQSWAP

Square root of Swap.

U(usize)

Abstract unitary gate with number of target qubits specified.

Methods

impl UnitaryGateType

pub fn into_gate_converter(
self,
num_controls: Option<usize>,
epsilon: f64,
ignore_global_phase: bool
) -> Box<dyn Converter<Input = Gate, Output = (Vec<QubitRef>, ArbData)>>