Function dqcsim::common::log::stdio::proxy_stdio

pub fn proxy_stdio(
    name: impl Into<String>, 
    stream: Box<dyn Read + Send>, 
    sender: Sender<LogRecord>, 
    level: Loglevel
) -> JoinHandle<()>

Forward standard i/o to log channel.

Spawns a thread which takes a readable stream and forwards lines as log records to the log thread until it matches EOF. The log record level is set to the level argument of the function.

Returns a thread::JoinHandle to the spawned thread.