Struct dqcsim::common::log::thread::LogThread

pub struct LogThread { /* fields omitted */ }

Methods

impl LogThread

pub fn spawn(
name: impl Into<String>,
proxy_level: LoglevelFilter,
stderr_level: LoglevelFilter,
callback: Option<LogCallback>,
tee_files: Vec<TeeFileConfiguration>
) -> Result<LogThread>

Spawn a [LogThread].

Returns [LogThread] instance if succesful. Also spawns a [LogProxy] in the current thread with the provided [name] and [proxy_level] as [LogLevelFilter].

Output to Standard Error can be enabled by settings the [stderr_level] above [LoglevelFilter::Off]. Output by invocatio of a callback function can be enabled by passing a [LogCallback] to [callback].

pub fn get_sender(&self) -> Sender<LogRecord>

pub fn get_ipc_sender(&self) -> IpcSender<LogRecord>