Module dqcsim::host::configuration

Configuration structures for the plugins and simulator.

Structs

PluginLogConfiguration
Configuration structure for the plugin logging system.

PluginProcessConfiguration
Represents the complete configuration for a plugin running in an external process.

PluginProcessFunctionalConfiguration
Structure describing the functional configuration of a plugin, i.e. the parameters that affect a plugin's behavior.

PluginProcessNonfunctionalConfiguration
Structure describing the NONfunctional configuration of a plugin, i.e. the parameters that only affect how the plugin represents its output.

PluginProcessSpecification
Plugin specification, consisting of the executable filename for the plugin and an optional script filename for it to execute for when the executable is an interpreter.

PluginThreadConfiguration
Represents the complete configuration for a plugin running in a local thread.

Seed
Represents a 64-bit random seed. Can be converted to and from a string as you would expect a seed to be. Also implements std::default::Default, which returns a random seed (from the random crate).

SimulatorConfiguration
The complete configuration for a DQCsim run.

Enums

EnvMod
Represents an environment variable modification.

ReproductionPathStyle
Represents the style for storing paths in a reproduction file.

StreamCaptureMode
Stream capture mode.

Timeout
Represents a timeout parameter, which may be infinite.

Traits

PluginConfiguration
Trait for types of configurations. The only thing that needs to be implemented is a way to instantiate a Plugin from it, and to return the log configuration of the Plugin.