Struct dqcsim::host::reproduction::PluginModification

pub struct PluginModification {
    pub name: String,
    pub verbosity: Option<LoglevelFilter>,
    pub tee_files: Vec<TeeFileConfiguration>,
    pub stdout_mode: Option<StreamCaptureMode>,
    pub stderr_mode: Option<StreamCaptureMode>,
    pub accept_timeout: Option<Timeout>,
    pub shutdown_timeout: Option<Timeout>,
}

Represents a nonfunctional configuration modification for a previously defined plugin.

This allows the nonfunctional configuration of the plugin to be modified. This is particularly important when DQCsim is running in --reproduce-exactly mode; it is of marginal use to reproduce a run exactly if you're not looking to get additional information from it by changing loglevels and such.

Fields
name: String

Name of the referenced plugin.

verbosity: Option<LoglevelFilter>

Specifies the verbosity of the messages sent to DQCsim. If this is None, the value of DQCsim's --plugin_level option should be used.

tee_files: Vec<TeeFileConfiguration>

Specifies the tee file configuration for this plugin.

stdout_mode: Option<StreamCaptureMode>

Specifies how the stdout stream of the plugin should be connected. None implies default.

stderr_mode: Option<StreamCaptureMode>

Specifies how the stderr stream of the plugin should be connected. None implies default.

accept_timeout: Option<Timeout>

Specifies the timeout for connecting to the plugin after it has been spawned.

shutdown_timeout: Option<Timeout>

Specifies the timeout for connecting to the plugin after it has been spawned.

Methods

impl PluginModification

pub fn apply(self, to: &mut Vec<PluginProcessConfiguration>) -> Result<()>

Applies this plugin modification to a plugin definition vector.

An error is returned if the referenced plugin cannot be found in the vector, otherwise Ok(()) is returned.