Struct dqcsim::host::simulation::Simulation

pub struct Simulation { /* fields omitted */ }

Simulation instance.

Methods

impl Simulation

pub fn new(
pipeline: Pipeline,
seed: Seed,
reproduction_log: Option<Reproduction>,
logger: &LogThread
) -> Result<Simulation>

Constructs a Simulation from a collection of PluginInstance and a random seed.

pub fn drop_plugins(&mut self)

Drains the plugin pipeline so their drop() implementations get called.

pub fn accelerator(&self) -> &Box<dyn Plugin>

pub fn accelerator_mut(&mut self) -> &mut Box<dyn Plugin>

pub fn yield_to_accelerator(&mut self) -> Result<()>

Yields to the accelerator.

The accelerator simulation runs until it blocks again. This is useful if you want an immediate response to an otherwise asynchronous call through the logging system or some communication channel outside of DQCsim's control.

This function silently returns immediately if no asynchronous data was pending or if the simulator is waiting for something that has not been sent yet.

pub fn arb(
&mut self,
name: impl AsRef<str>,
cmd: impl Into<ArbCmd>
) -> Result<ArbData>

Sends an ArbCmd message to one of the plugins, referenced by name.

ArbCmds are executed immediately after yielding to the simulator, so all pending asynchronous calls are flushed and executed before the ArbCmd.

pub fn arb_idx(
&mut self,
index: isize,
cmd: impl Into<ArbCmd>
) -> Result<ArbData>

Sends an ArbCmd message to one of the plugins, referenced by index.

The frontend always has index 0. 1 through N are used for the operators in front to back order (where N is the number of operators). The backend is at index N+1.

Python-style negative indices are supported. That is, -1 can be used to refer to the backend, -2 to the last operator, and so on.

ArbCmds are executed immediately after yielding to the simulator, so all pending asynchronous calls are flushed and executed before the ArbCmd.

pub fn get_metadata(&self, name: impl AsRef<str>) -> Result<&PluginMetadata>

Returns a reference to the metadata object belonging to the plugin referenced by instance name.

pub fn get_metadata_idx(&self, index: isize) -> Result<&PluginMetadata>

Returns a reference to the metadata object belonging to the plugin referenced by index.

pub fn write_reproduction_file(&self, filename: impl AsRef<Path>) -> Result<()>

Writes a the reproduction log to a file.