Module dqcsim::plugin::connection

Plugin to simulator connection wrapper.

The connection module provides a Connection wrapper, which is used by Plugin instances to wrap connection structures required for communication between Plugin and Simulator, and between Plugin and Plugin.

This module defines two wrapper enumerations, which wrap around protocol messages for both incoming (IncomingMessage) and outgoing (OutgoingMessage) channels.

Structs

Connection
Plugin to Simulator connection wrapper.

Enums

IncomingMessage
Incoming messages variants.

OutgoingMessage
Outgoing messages variants.