Struct dqcsim::plugin::connection::Connection

pub struct Connection { /* fields omitted */ }

Plugin to Simulator connection wrapper.

This provides a Plugin with the ability to communicate with both a Simulator instance and other upstream and downstream plugins.

Constructing a Connection instance should be the first thing a Plugin does. The Simulator server address string is passed as an argument to all Plugins started by a Simulator. The server address string can be used to construct a Connection instance.

After constructin of the Connection instance, the Plugin can respond to the initialization request from the Simulator.

Implementations

impl Connection

pub fn new(simulator: impl Into<String>) -> Result<Connection>

Construct a Connection wrapper instance.

The Connection wrapper attempts to connect to the Simulator using the address provided as argument. The required communication channels are generated and exchanged with the Simulator.

At this point the Connection wrapper can receive requests and send responses from and to the Simulator, however logging and upstream and downstream plugin connections are not yet available.

The first request the Simulator sends is always an initialization request, which should be handled with the init method.

pub fn connect_downstream(
&mut self,
downstream: impl Into<String>
) -> Result<()>

Connects to a downstream plugin.

pub fn serve_upstream(&mut self) -> Result<String>

Creates a one-shot server for an upstream plugin to connect to, returning the address. Call accept_upstream() to finish connecting.

pub fn accept_upstream(&mut self) -> Result<()>

Waits for an upstream plugin to connect to our one-shot server and establishes the connection. Call serve_upstream() first.

pub fn send(&self, message: OutgoingMessage) -> Result<()>

Send an OutgoingMessage.

Send an OutgoingMessage using the corresponding sender. Returns an error when the channel is closed, does not exist or when sending failed.

pub fn next_request(&mut self) -> Result<Option<IncomingMessage>>

Fetch next request from either the Simulator request channel or the upstream Plugin request channel.

Fails if either connection closed or if any connection is unexpectedly already closed. Returns Ok(None) if both request channels are closed. This method blocks until a new request is available.

pub fn next_downstream_request(&mut self) -> Result<Option<IncomingMessage>>

Fetch next downstream request.

Fails if the downstream connection is closed. This method blocks until a new request is available.