Module dqcsim::common::gates

Gate types and definitions.

The types defined here are provided to facilitate plugin developers. They are not to be confused with the Gate type used in gatestream protocol, and are not relied on in the core of DQCsim.

The following gate types are defined in this module:

UnitaryGateType: An abstract gate type. The variants in this type carry no additional parameters or information about the target qubits, with the exception of the UnitaryGateType::U variant that encodes an abstract unitary gate with the number of qubits involved specified.

UnboundUnitaryGate: An unbound gate type. The variants in this type specify all parameters to determine the behavior of the gate. However, these variants carry no information about the target qubits.

BoundUnitaryGate: A bound gate type. The variants in this type specify all parameters and qubits targets to determine the behaviour of the gate.

The BoundUnitaryGate can always be converted to an UnboundUnitaryGate that in turn can always be converted to a UnitaryGateType. UnboundUnitaryGate variants without additional parameters can be converted to their UnitaryGateType variants. BoundUnitaryGate instances can be converted to Gate instances.

Enums

BoundUnitaryGate
A bound gate type.

UnboundUnitaryGate
An unbound gate type.

UnitaryGateType
An abstract gate type.